🤖 Add EVENT constant
- click, touchstart event 명 상수 추가
- tag관련 상수 분리
