+++ src/Component/Utils/listenOutsideClick.ts
@@ -1,4 +1,5 @@
 import { RefObject } from 'react';
+import { EVENT } from '../../constants';
 
 export default function listenOutsideClick( //menuRef가 가르키는 dom요소(드롭다운)가 아닌 곳이 클릭되면 isOpen false로
   menuRef: RefObject<HTMLDivElement>,
@@ -7,7 +8,7 @@ export default function listenOutsideClick( //menuRef가 가르키는 dom요소(
   return () => {
     if (!menuRef.current) return;
 
-    ['click', 'touchstart'].forEach((e) => {
+    [EVENT.CLICK, EVENT.TOUCHSTART].forEach((e) => {
       document.addEventListener(e, (event) => {
         const curr = menuRef.current;
         const target = event.target;
