+++ src/constants/event.ts
@@ -1,0 +1,6 @@
+const EVENT = {
+  CLICK: 'click',
+  TOUCHSTART: 'touchstart',
+};
+
+export { EVENT };
